Prerelease Candidate 1
dzil configs updated to reflect my standard dzil layout.  Renaming the
scripts in the bin directory to have clearer, more appropriate names and
avoid collisions.  Restructure the timing data so it can be easily
indexed in ElasticSearch.  Convert the config file to an optional
parameter so the utilities run cleanly from the the dist.