Add static context to add k/v pairs to every log event.
Add a special "double star" matcher to match every event.
Add a "SuppressWarnings" variable to the contexts.  If a context wishes
to be silently ignored, it can set "our $SuppressWarnings".

Use both of these features in the static context.  The advantage is one
less subroutine dispatch if the static context isn't configured.