Fix the eris-es-indexer.pl
The config was pointing to the wrong depth in the hash. Ensure when the
config is passed from the commandline, eris::schemas are instantiated
correctly.  Fix the mapping for the geo_point field in the geoip
mapping.

Add the eris::log::context::iptables to parse iptables logs into the
indexes.